KVM-A15: boot-wrapper: fix non-secure SMP boot
The boot wrapper suffers from a few problems when it comes to
setup the secondary CPUs, as only the boot CPU hits the non-secure
setup path.

Move the non-secure setup to the common path, taking care of only
configuring GICD_IGROUPR0 on secondary CPUs.

Also set bit 18 in NSACR to allow the SMP bit to be set in the
Auxiliary Control Register (A15 specific).

Tested on A15 model v6.1.70.

Cc: Christopher Dall <cdall@cs.columbia.edu>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>