Makefile: Introduced super-simple config file
Config files are named config.mk. This file is added to .gitignore,
but a default config file is supplied in config-default.mk.

The default config file creates kernel command boot lines for NFS boots
based on a script obtaining the host IP addres. Naturally users can
change this to a static IP or another script if they wish.

The config file lets users select a system and other config options
depend on this overall setting.

I am no expert on Makefiles, so there could be better ways to accomplish
these things, but I think this suffices for now.


Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>