config-default.mk: Explicitly pass -marm when we want to build ARM mode
Explicitly pass -marm in the CFLAGS when we want to build in ARM mode,
because some compilers will default to Thumb. This has not previously been
a problem in practice because the assembler will compile in ARM mode
anyway unless we told it to use Thumb mode with -Wa,-mthumb. However it
does mean that we end up building boot.S with a cpp that thinks it is
compiling for Thumb (and thus defines __thumb__) but an assembler that
generates ARM code. Passing -marm explicitly avoids this mismatch and
allows us to use the cpp preprocessor defines without fear.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>