boot-wrapper: Support reading kernel/initrd via semihosting
Implement support in the boot-wrapper for reading our kernel,
initrd and command line arguments using the semihosting ABI
provided by a Fast Model. We build a second AXF file
linux-system-semi.axf which does not depend on or bake in
a kernel binary. Instead when you start the model with it
you pass the model a parameter like:
  -C cluster.cpu0.semihosting-cmd_line="--kernel /path/to/uImage
  --initrd /path/to/initrd -- console=ttyAMA0 mem=512M ..."

The boot wrapper will interpret the --kernel and --initrd
arguments as taking an option which is a (host system)
path, and pass the tail of the string following '--' to
the kernel as its command line.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>