Fix linux-system-semi.axf build failure from clean
We were failing to build linux-system-semi.axf in a clean tree
because even if SEMIHOSTING was defined we still had a spurious
INPUT(boot.o), which is harmless if boot.o exists but fails
if it does not.

In fact the INPUT() commands for the .o files are unnecessary
because ld will pull in the correct .o files because the SECTIONS
command has a non-wildcard reference to them. So just drop the
unneeded INPUT() commands rather than adding more ifdefs.
(The INPUT() commands for the binary blobs are still required.)

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>