boot-wrapper: tidy the Makefile to help with packaging.
Add and modify Makefile rules so that it can be used in a Debian
package.  Add an explicit 'distclean' target.  Add a 'semi' target
that only builds the semihosting binary.  Add stub commands to 'all'
and 'semi' so the call doesn't fall through to the default kernel
rule.  Update 'clean' to also clean the semihosting files.

Signed-off-by: Michael Hope <michael.hope@linaro.org>
Signed-off-by: Rusty Russell <rusty.russell@linaro.org>