Makefile: Use .PHONY target, not dummy commands; tidy clean list
There's no need to use a dummy 'true' command to suppress the
default 'try kernel makefile' target for targets with no actual
rules -- you can just list as in .PHONY instead.

Tidy the list of files to remove on clean to remove a duplicate
and use $(KERNEL) instead of uImage.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>