bootwrapper: Import C semihosting code
Import the reimplementation in C of the semihosting support
written by Dave Martin for the ARM big.LITTLE bootwrapper.
The main benefit of this switchover (apart from only having
one implementation rather than two) is the fdt support it
includes.

This commit only imports the sources, it doesn't compile them.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>