bootwrapper: Use local definition of string.h
The semihosting and FDT code makes use of libc style string
functions implemented in our string.c, however it relies on the system
providing the string.h header file.

This causes problems on toolchains that don't provide these headers,
like Android toolchains, and it also means that we include declaration
for functions which aren't implemented in the bootwrapper.

Resolve this by providing our own string.h which declares only
the functions we implement and add the base directory to the
include path so this header is found.

Signed-off-by: Jon Medhurst <tixy@linaro.org>
Signed-off-by: Peter Maydell <peter.maydell@linaro.org>