bootwrapper: Delay switch to Hyp mode until kernel entry
The bootwrapper is really doubling as firmware, so it doesn't make
sense for it to drop out of the Secure World before getting a
chance to parse its parameters and configuration.

Instead, it is better to delay switching to the Normal World for
as long as possible so that we have a chance to do any required
firmware-level configuration in the Secure World first.

Signed-off-by: Dave Martin <dave.martin@linaro.org>
[PMM: moved enter_hyp for secondary CPUs to just before the WFE
 loop to avoid making the loop code that needs relocation bigger]
Signed-off-by: Peter Maydell <peter.maydell@linaro.org>