bootwrapper: Initialise CCI device if found in the fdt
The A15xA7 models simulate a Cache Coherent Interconnect (CCI) and this
needs to be initialised correctly for Linux to boot.

To perform this initiation we add the new function configure_from_fdt()
which will look in the fdt for devices to initialise. In this first case
we look for the CCI node and if found then setup this device.

Signed-off-by: Jon Medhurst <tixy@linaro.org>
Acked-by: Dave Martin <dave.martin@linaro.org>
Signed-off-by: Peter Maydell <peter.maydell@linaro.org>